+++ spec/unit/profile_parser_spec.rb
@@ -2,13 +2,9 @@ require 'spec_helper'
 
 require "#{PROJECT_ROOT}/puppet-profile-parser.rb"
 
-describe PuppetProfiler do
-  it { is_expected.to be_a(Module) }
-end
-
-describe PuppetProfiler::LogParser do
+describe PuppetProfileParser::LogParser do
   describe 'ISO 8601 timestamp parser' do
-    subject { PuppetProfiler::LogParser::ISO_8601 }
+    subject { PuppetProfileParser::LogParser::ISO_8601 }
 
     it 'matches timestamps produced by the logback %date pattern' do
       expect(subject.match('2017-09-18 12:01:43,344')).to be_a(MatchData)
@@ -16,7 +12,7 @@ describe PuppetProfiler::LogParser do
   end
 
   describe 'Puppet Server default logback layout parser' do
-    subject { PuppetProfiler::LogParser::DEFAULT_PARSER }
+    subject { PuppetProfileParser::LogParser::DEFAULT_PARSER }
 
     it 'matches PROFILE output formatted with the layout' do
       result = subject.match("2018-02-18 18:43:53,501 INFO  [qtp1732817189-1224] [puppetserver] Puppet PROFILE [39776666] 1 Processed request GET /puppet/v3/node/pe-201734-master.puppetdebug.vlan: took 0.1640 seconds\n")